requirements.txt
opencv-python
numpy
pandas
Pillow
scikit-learn
tqdm
pyyaml
python-dotenv

torch
torchvision
timm
albumentations
tensorboard

ultralytics

fal-client
lpips
git+https://github.com/openai/CLIP.git

streamlit
matplotlib
seaborn
